+++ packages/common/src/interfaces/type-helper.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * Makes the object deep partial.
+ */
+export type DeepPartial<T> = {
+  [P in keyof T]?: DeepPartial<T[P]>
+}