+++ packages/common/src/lib/config/config.constants.ts
@@ -1,0 +1,6 @@
+export enum ConfigEnvKeys {
+  NAME = '__name',
+  PARSER = '__format',
+  ELEMENT = '__element',
+  ELEMENT_REPLACER = '${i}'
+}