+++ packages/common/src/lib/config/config.interface.ts
@@ -1,15 +1,20 @@
 import type { LogLevels } from '@utils/logger'
 
-export interface GlobalConfig {
+export interface GlobalConfigInit {
   logLevel: LogLevels
-
   ci: boolean
-
-  isVerbose: boolean
-
-  isDebug: boolean
-
-  isSilent: boolean
-
   json: boolean
+}
+
+export interface GlobalConfig extends GlobalConfigInit {
+  isVerbose: boolean
+  isDebug: boolean
+  isSilent: boolean
+}
+
+export interface ConfigIterator {
+  key: (string | number)[]
+  env: string
+  parser?: string
+  extensions?: ConfigIterator[]
 }